Move ws server
